refactor: reorganize jsonapi-rules into multiple files
